style(*): change prettier's max line length to 120

+++ src/localized-content/localized-content.service.ts
@@ -2,19 +2,14 @@ import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository, EntityManager } from "typeorm";
 
-import {
-  LocalizedContentEntity,
-  LocalizedContentType
-} from "./localized-content.entity";
+import { LocalizedContentEntity, LocalizedContentType } from "./localized-content.entity";
 import { Locale } from "@/common/locale.type";
 
 @Injectable()
 export class LocalizedContentService {
   constructor(
     @InjectRepository(LocalizedContentEntity)
-    private readonly localizedContentRepository: Repository<
-      LocalizedContentEntity
-    >
+    private readonly localizedContentRepository: Repository<LocalizedContentEntity>
   ) {}
 
   async createOrUpdate(
@@ -56,16 +51,11 @@ export class LocalizedContentService {
 
     if (locale) match.locale = locale;
 
-    if (transactionalEntityManager)
-      await transactionalEntityManager.delete(LocalizedContentEntity, match);
+    if (transactionalEntityManager) await transactionalEntityManager.delete(LocalizedContentEntity, match);
     else await this.localizedContentRepository.delete(match);
   }
 
-  async get(
-    objectId: number,
-    type: LocalizedContentType,
-    locale: Locale
-  ): Promise<string> {
+  async get(objectId: number, type: LocalizedContentType, locale: Locale): Promise<string> {
     const localizedContent = await this.localizedContentRepository.findOne({
       objectId: objectId,
       type: type,
@@ -75,39 +65,27 @@ export class LocalizedContentService {
     return localizedContent ? localizedContent.data : null;
   }
 
-  async getOfAllLocales(
-    objectId: number,
-    type: LocalizedContentType
-  ): Promise<Partial<Record<Locale, string>>> {
+  async getOfAllLocales(objectId: number, type: LocalizedContentType): Promise<Partial<Record<Locale, string>>> {
     const localizedContents = await this.localizedContentRepository.find({
       objectId: objectId,
       type: type
     });
 
     let result: Partial<Record<Locale, string>> = {};
-    for (const localizedContent of localizedContents)
-      result[localizedContent.locale] = localizedContent.data;
+    for (const localizedContent of localizedContents) result[localizedContent.locale] = localizedContent.data;
     return result;
   }
 
-  async getOfAnyLocale(
-    objectId: number,
-    type: LocalizedContentType
-  ): Promise<[Locale, string]> {
+  async getOfAnyLocale(objectId: number, type: LocalizedContentType): Promise<[Locale, string]> {
     const localizedContent = await this.localizedContentRepository.findOne({
       objectId: objectId,
       type: type
     });
 
-    return localizedContent
-      ? [localizedContent.locale, localizedContent.data]
-      : null;
+    return localizedContent ? [localizedContent.locale, localizedContent.data] : null;
   }
 
-  async countLocales(
-    objectId: number,
-    type: LocalizedContentType
-  ): Promise<number> {
+  async countLocales(objectId: number, type: LocalizedContentType): Promise<number> {
     return await this.localizedContentRepository.count({
       objectId: objectId,
       type: type
